Track and Queue sequencers semi working with CausticLive

+++ CausticToolkit/src/main/java/com/teotigraphix/caustk/sequencer/track/TrackPhrase.java
@@ -242,7 +242,7 @@ public class TrackPhrase implements ISerialize {
                 new OnTrackSequencerPropertyChange(PropertyChangeKind.EditMeasure, this));
     }
 
-    public void onBeatChange(int beat) {
+    public void onBeatChange(float beat) {
         localBeat = (int)toLocalBeat(beat, getLength());
 
         float fullMeasure = beat / 4;
